demos/ai-architecture-simulator/requirements.txt
streamlit>=1.31.0
openai>=1.0.0
plotly>=5.18.0
python-dotenv>=1.0.0
